create review method adjustment rating review brand model service

+++ server/src/properties/entities/tyre-brand.model.ts
@@ -7,7 +7,7 @@ import { ReviewTyres } from "src/reviews/entities/review-tyres.model";
 @Table({tableName: 'tyre_brand' , updatedAt: false, createdAt: false})
 export class TyreBrand extends Model<TyreBrand, TyreBrandConfigAttr> {
 
-    @Column({type: DataType.BIGINT, unique: true, allowNull: false, primaryKey: true, autoIncrement:false})
+    @Column({type: DataType.INTEGER, unique: true, allowNull: false, primaryKey: true, autoIncrement:false})
     id: number;
    
     @Column({type: DataType.STRING, unique: true, allowNull: false})
